Initial support to manage branch protections
This adds basic support to manage branch protections via git. We'll
first start testing on ansible-network/sandbox then move to other
projects.

The idea, is we drop down a .config file, if the settings we want to
enforce for each branch, on each project.  We can also include a
top-level common file, for shared projects if needed.

Signed-off-by: Paul Belanger <pabelanger@redhat.com>